Support dequoting PostgreSQL 9.0 "hex" BYTEA output format
In PostgreSQL 9.0, the default BYTEA output format has changed from
octal escapes for non-printable-ascii characters to encoding the whole
value in hex (with a leading "\x" to distinguish the two).

This also removes papering-over of the lack of dequoting from
t/03dbmethod.t and extends t/06bytea.t to cover both output formats.