Test that the Pg server agrees with us about the lengths of input strings.
This should stop us missing bugs where we send a string encoded in some
bogus form, and accurately reverse the transformation on return. If we
happen to have forwards and backwards bogosity exactly matched, then the
Perl would think that it was communicating correctly because it got the same
answer back, and wouldn't be aware of the mistake.