Make get_info much more efficient and slightly simpler.
Previously on every call get_info copied a large hash into another hash,
then added more keys to the new hash, then discarded the new hash.
I've moved the hash out from the method and made it static.
There's no longer any hash copying.

Also there's a long elsif chain to handle special cases (which has a
poor-failure mode). I've added two new general mechanisms to remove the
need for many of the special cases: A code ref in the hash is called
with the dbh and the return value returned.  A scalar ref is treated as
a ref to an SQL statement which is executed and the first value returned.

The same mechanisms should be extended to the other special cases but I
stopped here to keep the patch reasonably clear.