Fix parsing of PostgreSQL versions >= 10 on Debian/Ubuntu
Debian and Ubuntu configure postgres >= 10 using --with-extra-version to
add the package version, so that pg_config --version outputs
e.g. "PostgreSQL 10.6 (Ubuntu 10.6-1.pgdg18.10+1)"

The regex for three-part versions was missing anchoring and had a "."
instead of "\.", so it matched the "10.6-1" part and set PGLIBVERSION to
100601 instead of 100006.

In passing, move the beta/devel/rc comment to the right place and add a
comment about the tree-part version numbers.