Inline no-longer-conditional catalog query fragments
Some bits of catalog queries needed version-specific variants, but when
support for the versions that necessitated them was removed, the
variables were left in place.  This removes these now-unnecessary
variables and puts the SQL fragments back in the query.

The pg_get_expr() call does not need to be conditional, it exists in 7.4
too: https://www.postgresql.org/docs/7.4/functions-misc.html