Fetch statistics_info completely in SQL
This combines the table and index stats in a single UNION ALL query,
avoiding repetition of the column lists and the extra buffering via
DBD::Sponge.  It also reduces the the risk of mismatches betwen the
two queries, since SQL requres the column types of a UNION to match.

This changes the return value from undef to an empty result set when
the specified table doesn't exist and the user not limiting the
results to unique indexes.  The new behaviour mathces the behaviour
when requesting only unique indexes, and the behaviour of other DBDs.