Speed up quote_name using binary search
The current implementation of quote_name calls is_keyword, which
runs a linear scan of every postgres keyword looking for a match.
Since most names passed to quote_name will not be a keyword, this
is likely to scan the full list every time.

A much more performant strategy is to make a single character
comparison that divides the list in half repeatedly until there is
only one possible option, then run a single strcmp() on that.

This executes 10x as fast, and also compiles into smaller machine
code (despite looking like a lot more instructions) because it has
only one function call, and a function call takes more instructions
than a single-character comparison + jump.