Unit test for is_keyword
Since the implementation of is_keyword is now so much more
complicated, it deserves its own unit test to ensure that all
the keywords still get detected.

To do this, I had to expose is_keyword to Perl, which I did as
DBD::Pg::db::_is_keyword.  This could maybe be useful for other
purposes down the line.