Reconcile problems between 01keywords.t and 99_lint.t
99_lint.t expected all lines containing Test::More verb 'like'
to be a specific format, but detected the 'like' in the middle of
the postgres keywords.  I updated the generator for the unit test
to put quotes around the keywords, and then updated 99_lint.t to
ignore things that look like quoted strings.

I also updated 01keywords.t to use single-quoted strings.